moved http calls out of mutations in edit.ts, closes TLI-38
